vsphere: fix incorrect PoweredOn conditional
Previously it returned true when the VM was _not_ powered on, which is the
opposite of what we want. This caused the janitor to try to destroy all VMs that
were powered on (but luckily it thought they were powered off, so it tried to
destroy them without powering them off first, causing vSphere to error and the
VMs were left alone).